Records (#1)
* use records for DTOs

* fix mapping issues

* fix nameSurname issue

* fix issue

* fix null ignore problems

* fix mapping problem

* complete mappings

---------

+++ src/Presentation/ViewModels/Product/AddEditProductViewModel.cs
@@ -21,6 +21,6 @@ namespace Presentation.ViewModels.Product
 
         [Required(ErrorMessage = ValidationErrorCode.RequiredField)]
         [Display(Name = "STOCK_QUANTITY")]
-        public decimal? StockQuantity { get; set; }
+        public int? StockQuantity { get; set; }
     }
 }
